Add RateLimiter end-to-end test

+++ test/helpers_aeron.jl
@@ -138,7 +138,7 @@ Poll until an attach response with the given correlation id is received.
 function await_attach!(
     state::DriverClientState,
     correlation_id::Int64;
-    timeout_ns::UInt64 = 5_000_000_000,
+    timeout_ns::UInt64 = UInt64(5_000_000_000),
 )
     now_ns = time_ns()
     deadline = now_ns + timeout_ns
